refactor: split custom action helpers
